+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [1.5.1]
   * add [Progress.circular] [Progress.linear]
   * ExtensionString add insert void
+  * add [MergeText]  [BasisTextStyle] 
+  * modify [RichSpan] to [RichTextSpan]
 ## [1.5.0]
   * change Tools to Ts.
   * change GlobalMaterial to GlobalWidgetsApp.